Server Summary Fix

+++ Shared.Rcl/Commands/Servers/Nics/ServerNicAddCommand.cs
@@ -10,7 +10,7 @@ public class ServerNicAddSettings : ServerNameSettings
 {
     [CommandOption("--type <TYPE>")] public string Type { get; set; }
 
-    [CommandOption("--speed <SPEED>")] public int Speed { get; set; }
+    [CommandOption("--speed <SPEED>")] public double Speed { get; set; }
 
     [CommandOption("--ports <PORTS>")] public int Ports { get; set; }
 }
